Fix line and copy

+++ src/VisionSoilAnalyzer/SoilPlot/DrawFigure.cpp
@@ -45,7 +45,7 @@ namespace SoilPlot
 			{
 				cv::Vec4b *F = &Figure.at<cv::Vec4b>(cv::Point(i, j));
 				cv::Vec4b *G = &workImg.at<cv::Vec4b>(cv::Point(i, j));
-				float ratio = (*F)[3] / 255;
+				float ratio = (float)(*F)[3] / 255;
 				for (uint32_t k = 0; k < 3; k++)
 				{
 					int newVal = (*G)[k] * (1 - ratio) + (*F)[k] * ratio;
